fix: [Checkboxes] Interaction on Storybook overview page was not updating the checkbox state
