+++ src/components/Checkbox/Checkbox.test.tsx
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom';
 import { act, render, screen } from '@testing-library/react';
-import { useState } from 'react';
 import { Checkbox } from './Checkbox';
+import { CheckboxTestWrapper } from './Checkbox.utils';
 
 const id = 'default';
 const label = 'this is a label';
@@ -12,34 +12,6 @@ const attributeClass = 'class';
 const attributeTitle = 'title';
 
 const defaultProps = { id, label, 'data-testid': testId };
-
-/*
- * Helper for controlled component state.
- * ARIA attributes won't updated without the updated onClick state being fed back into the component.
- */
-const CheckboxWrapper = ({
-  onChange,
-  ...others
-}: typeof Checkbox): JSX.Element => {
-  const [checked, setChecked] = useState(false);
-
-  const onWrapperChange = (
-    event: React.ChangeEvent<HTMLInputElement>
-  ): void => {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-    onChange?.(event);
-    setChecked(event.target.checked);
-  };
-
-  return (
-    <Checkbox
-      {...defaultProps}
-      {...others}
-      checked={checked}
-      onChange={onWrapperChange}
-    />
-  );
-};
 
 describe('Checkbox', () => {
   it('Propagates additional HTML properties to main component element', () => {
@@ -55,7 +27,7 @@ describe('Checkbox', () => {
   it('Calls the provided onChange handler', async () => {
     const onChange = vi.fn();
 
-    render(<CheckboxWrapper {...defaultProps} onChange={onChange} />);
+    render(<CheckboxTestWrapper {...defaultProps} onChange={onChange} />);
 
     const checkbox = await screen.findByRole(/checkbox/i);
 
@@ -77,7 +49,7 @@ describe('Checkbox', () => {
   it('Renders helper text that toggles checkbox when clicked', async () => {
     const helperText = 'This is optional helper text for the checkbox';
 
-    render(<CheckboxWrapper {...defaultProps} helperText={helperText} />);
+    render(<CheckboxTestWrapper {...defaultProps} helperText={helperText} />);
 
     const checkbox = await screen.findByRole(/checkbox/i);
     expect(checkbox.getAttribute(attributeAria)).toMatch('false');
